Fixes and upgrades, add Freeze Over - Add FreezeOver class and spell - Fix Thunderhorse so players can't steal Thunderhorse's saddle - Fixed MagicWand to have static methods - Added multiple modes to Wizard Shovel and Wizard Pick - Added mode-toggling to Wizard Shovel and Wizard Pick - Added BlockBoxer class with static functions to help create lists of Blocks -

+++ me/mistergone/AWizardDidIt/helpers/MagicCommands.java
@@ -46,7 +46,7 @@ public class MagicCommands implements CommandExecutor {
                         WizardPlayer wizardPlayer = getWizardry().getWizardPlayer( p.getUniqueId() );
                         double wizardPower = wizardPlayer.getWizardPower();
                         getWizardry().getWizardPlayer( p.getUniqueId() ).showWizardBar();
-                        p.sendMessage( "You have " + String.valueOf( Math.floor( wizardPower * 1000 ) )
+                        p.sendMessage( "You have " + String.valueOf( (int) Math.floor( wizardPower * 1000 ) )
                                 +  " points of Wizard Power.");
                         if ( wizardPlayer.getSpells().size() > 0 ) {
                             for ( String spell: wizardPlayer.getSpells() ) {
